web: split organization and application layouts
